refactor: 모듈 간 중복 코드 줄이기 (#1)
