+++ breaditnow-customer-api/build.gradle
@@ -23,3 +23,11 @@ dependencies {
 test {
     useJUnitPlatform()
 }
+
+bootJar{
+    enabled=false
+}
+
+jar{
+    enabled=true
+}
