--- breaditnow-customer-api/src/main/resources/application.yml
@@ -1,9 +1,0 @@
-server:
-  port: 9000
-
-spring:
-  application:
-    name: breaditnow-customer-api
-
-  profiles:
-    active: local
