+++ breaditnow-mysql/src/main/java/com/breaditnow/mysql/BreaditnowMysqlApplication.java
@@ -1,4 +1,4 @@
-package com.breaditnow.mysql.repository;
+package com.breaditnow.mysql;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
