--- breaditnow-mysql/src/main/resources/application-local.yml
@@ -1,29 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: local
-
-  docker:
-    compose:
-      enabled: true
-      file: ./docker/docker-compose.yml
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL8Dialect
-        format_sql: true
-        show_sql: true
-    open-in-view: false
-
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/breaditnow?serverTimezone=Asia/Seoul&useSSL=false
-    username: root
-    password:
-
-  redis:
-    host: localhost
-    port: 6379
