+++ breaditnow-mysql/src/main/resources/application-mysql-local.yml
@@ -1,7 +1,8 @@
 spring:
-  config:
-    activate:
-      on-profile: local
+  docker:
+    compose:
+      enabled: true
+      file: ./docker/docker-compose.yml
 
   jpa:
     hibernate:
@@ -17,7 +18,7 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/breaditnow?serverTimezone=Asia/Seoul&useSSL=false
     username: root
-    password: root
+    password:
 
   redis:
     host: localhost
