--- breaditnow-owner-api/src/main/resources/application.yml
@@ -1,9 +1,0 @@
-server:
-  port: 9001
-
-spring:
-  application:
-    name: breaditnow-owner-api
-
-  profiles:
-    active: local
