+++ common/build.gradle
@@ -2,18 +2,38 @@ plugins {
     id 'java'
 }
 
-group = 'com.breaditnow'
-version = 'unspecified'
-
-repositories {
-    mavenCentral()
+allprojects {
+    repositories {
+        mavenCentral()
+    }
 }
 
-dependencies {
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
+subprojects {
+    group = 'com.breaditnow'
+    version = '0.0.1-SNAPSHOT'
 
-    implementation 'org.springframework.boot:spring-boot-starter-web'
+    apply plugin: 'java'
+    apply plugin: 'org.springframework.boot'
+    apply plugin: 'io.spring.dependency-management'
+    apply plugin: 'java-library'
+
+    dependencies {
+        implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+        implementation 'org.springframework.boot:spring-boot-starter'
+        testImplementation 'org.springframework.boot:spring-boot-starter-test'
+        testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+        compileOnly 'org.projectlombok:lombok'
+        annotationProcessor 'org.projectlombok:lombok'
+
+        implementation 'com.mysql:mysql-connector-j'
+
+        //testcontainer(root)
+        testImplementation "org.testcontainers:testcontainers:1.17.6"
+        testImplementation "org.testcontainers:junit-jupiter:1.17.6"
+    }
 }
 
 test {
